Corrected bugs in the conversion functions
The functions all assumed that they were working on single file path 
results, instead of being handed a list of file paths. Given that, the
original tests were passing, based on previous assumptions, but the 
code’s functionality was wrong, as the core assumptions were incorrect.
With these changes, the code now assumes that it’s given a list of 
file paths.